PricingPlanDiscount

+++ app/src/main/java/com/nowfloats/Store/Model/ProductPaymentModel.java
@@ -14,13 +14,13 @@ public class ProductPaymentModel {
     private String description;
     @SerializedName("discount")
     @Expose
-    private Integer discount;
+    private Double discount;
     @SerializedName("name")
     @Expose
     private String name;
     @SerializedName("price")
     @Expose
-    private Integer price;
+    private Double price;
     @SerializedName("product_id")
     @Expose
     private String productId;
@@ -49,11 +49,11 @@ public class ProductPaymentModel {
         this.description = description;
     }
 
-    public Integer getDiscount() {
+    public Double getDiscount() {
         return discount;
     }
 
-    public void setDiscount(Integer discount) {
+    public void setDiscount(Double discount) {
         this.discount = discount;
     }
 
@@ -65,11 +65,11 @@ public class ProductPaymentModel {
         this.name = name;
     }
 
-    public Integer getPrice() {
+    public Double getPrice() {
         return price;
     }
 
-    public void setPrice(Integer price) {
+    public void setPrice(Double price) {
         this.price = price;
     }
